Refactor AudioFile domain folder hierarchy
