+++ includes/IO/AudioFile/lib/impl/AudioFileRepository_libsndfile.cpp
@@ -22,7 +22,7 @@
 #include <string>
 #include <variant>
 
-#include "../AudioFile.h"
+#include "../../AudioFile.h"
 #include "AudioFileRepository_libsndfile.h"
 
 #include "filesystem_Delete.h"
