+++ includes/IO/AudioFile/lib/impl/libsndfile_Append.cpp
@@ -19,43 +19,47 @@
  */
 
 #include <stdexcept>
-#include <vector>
 
 #include "sndfile.h"
 
-#include "../AudioFile.h"
-#include "../AudioFileWriteSettings.h"
+#include "../../AudioFile.h"
+#include "../../AudioFileWriteSettings.h"
 #include "libsndfile__Util.h"
 
 namespace Clover::IO::AudioFile::impl {
 
-void libsndfile_Write(const std::string &path,
-                      const WriteSettingsPcm &writeSettings,
-                      const AudioFile &audioFile) {
-  SF_INFO sfinfo;
-  sfinfo.samplerate = writeSettings.sampleRate;
-  sfinfo.channels = audioFile.channelCount;
-  sfinfo.format = getWriteSettingsFormat(writeSettings);
+void libsndfile_Append(const std::string &path,
+                       const WriteSettingsPcm &writeSettings,
+                       const AudioFile &audioFile) {
 
-  SNDFILE *file = sf_open(path.c_str(), SFM_WRITE, &sfinfo);
+  SF_INFO sfinfo;
+  SNDFILE *file = sf_open(path.c_str(), SFM_RDWR, &sfinfo);
+
   throwIfFails(file, sf_error(file));
 
+  bool sampleRatesMismatch = sfinfo.samplerate != audioFile.sampleRateHz;
+  if (sampleRatesMismatch) {
+    sf_close(file);
+    throw std::runtime_error(
+        "Incompatible sample rate when appending audio to file: path=[" +
+        std::string(path) + "]");
+  }
+
+  bool channelCountMismatch = sfinfo.channels != audioFile.channelCount;
+  if (channelCountMismatch) {
+    sf_close(file);
+    throw std::runtime_error(
+        "Incompatible channel count when appending audio to file: path=[" +
+        std::string(path) + "]");
+  }
+
+  sf_seek(file, 0, SEEK_END);
   sf_count_t count = sf_write_float(file, audioFile.audioData.data(),
                                     audioFile.audioData.size());
 
   if (count != static_cast<sf_count_t>(audioFile.audioData.size())) {
     throwIfFails(file, sf_error(file));
   }
-
-  std::vector<SF_CUE_POINT> cuePoints;
-  cuePoints.reserve(audioFile.cuePoints.size());
-  for (auto &point : audioFile.cuePoints) {
-    SF_CUE_POINT cuePoint;
-    cuePoint.position = point;
-    cuePoints.push_back(cuePoint);
-  }
-  sf_command(file, SFC_SET_CUE, cuePoints.data(),
-             cuePoints.size() * sizeof(SF_CUE_POINT));
 
   sf_write_sync(file);
   throwIfFails(file, sf_close(file));
