+++ includes/IO/AudioFile/lib/impl/libsndfile_Read.cpp
@@ -22,47 +22,43 @@
 
 #include "sndfile.h"
 
-#include "../AudioFile.h"
-#include "../AudioFileWriteSettings.h"
+#include "../../AudioFile.h"
 #include "libsndfile__Util.h"
 
 namespace Clover::IO::AudioFile::impl {
 
-void libsndfile_Append(const std::string &path,
-                       const WriteSettingsPcm &writeSettings,
-                       const AudioFile &audioFile) {
-
+AudioFile libsndfile_Read(const std::string &path) {
+  AudioFile audioFile;
   SF_INFO sfinfo;
-  SNDFILE *file = sf_open(path.c_str(), SFM_RDWR, &sfinfo);
-
+  SNDFILE *file = sf_open(path.c_str(), SFM_READ, &sfinfo);
   throwIfFails(file, sf_error(file));
 
-  bool sampleRatesMismatch = sfinfo.samplerate != audioFile.sampleRateHz;
-  if (sampleRatesMismatch) {
-    sf_close(file);
-    throw std::runtime_error(
-        "Incompatible sample rate when appending audio to file: path=[" +
-        std::string(path) + "]");
-  }
+  audioFile.sampleRateHz = sfinfo.samplerate;
+  audioFile.channelCount = sfinfo.channels;
 
-  bool channelCountMismatch = sfinfo.channels != audioFile.channelCount;
-  if (channelCountMismatch) {
-    sf_close(file);
-    throw std::runtime_error(
-        "Incompatible channel count when appending audio to file: path=[" +
-        std::string(path) + "]");
-  }
-
-  sf_seek(file, 0, SEEK_END);
-  sf_count_t count = sf_write_float(file, audioFile.audioData.data(),
-                                    audioFile.audioData.size());
+  audioFile.audioData.resize(sfinfo.frames * sfinfo.channels);
+  sf_count_t count = sf_read_float(file, audioFile.audioData.data(),
+                                   audioFile.audioData.size());
 
   if (count != static_cast<sf_count_t>(audioFile.audioData.size())) {
     throwIfFails(file, sf_error(file));
   }
 
-  sf_write_sync(file);
-  throwIfFails(file, sf_close(file));
+  SF_CUES cues;
+  if (sf_command(file, SFC_GET_CUE, &cues, sizeof(cues)) == SF_TRUE) {
+    for (unsigned i = 0; i < cues.cue_count; ++i) {
+      audioFile.cuePoints.push_back(cues.cue_points[i].sample_offset);
+    }
+  }
+
+  if (sf_close(file) != 0) {
+    int err = sf_error(file);
+    if (err != SF_ERR_NO_ERROR) {
+      throw std::runtime_error(sf_strerror(file));
+    }
+  }
+
+  return audioFile;
 }
 
 } // namespace Clover::IO::AudioFile::impl
