+++ test/functional/IO/AudioFileRepository.cc
@@ -24,8 +24,8 @@
 
 #include "Algo/Wavetable/WavetableOscillatorMono.h"
 #include "IO/AudioFile/AudioFile.h"
-#include "IO/AudioFile/AudioFileRepository.h"
 #include "IO/AudioFile/AudioFileWriteSettings.h"
+#include "IO/AudioFile/lib/AudioFileRepository.h"
 
 #include "./AudioFileRepository/AudioFileRepository_util.h"
